Enable 'isnan' implementation in lib/msun/src/s_isnan.c
In FreeBSD, 'isnan' is defined in libc.so or, in case of static
build, in lib/msun/src/s_isnan.c. This fix enable it's implementation
in lib/msun/src/s_isnan.c if __CRYSTAX__ macro defined.
This way we ensure we always have 'isnan' included to libcrystax
(to both static and shared versions).

Signed-off-by: Dmitry Moskalchuk <dm@crystax.net>